Add support for also hashing link and network layer information
This commit adds two flags to can-i-use-afpacket-fanout:

-includenetworklayer includes information about the link and network
layers into the hash that is used to determine if packets end up on the
correct thread.

This can be useful if you have, e.g., MPLS wrapped packets on your
network. If you see packets twice, once encapsulated in MPLS and once
outside of MPLS, can-I-use-afpacket will probably complain about
mismatched packets with this option, because it only looks at the IP
addresses and ports.

-dumpnetworklayerinformation outputs information about the number of
flows that were seen with each link and networklayer on each thread.
This can be useful to diagnose if a single thread is being overloaded,
e.g. with all MPLS packets.

Example output:

2021/04/28 16:57:04  - worker=11 flows=840244
2021/04/28 16:57:04    - protocol=Ethernet, IPv4, IPv6, ICMPv6, ICMPv6Echo flows=1
2021/04/28 16:57:04    - protocol=Ethernet, IPv4 flows=14954
2021/04/28 16:57:04    - protocol=Ethernet, MPLS, IPv4, ICMPv4, Payload flows=22907
2021/04/28 16:57:04    - protocol=Ethernet, MPLS, IPv4, Fragment flows=41